Removed PlayerController. Updated input method. Fixed jetpack. Fixed player collisions
