+++ Assembly-CSharp.csproj
@@ -63,7 +63,6 @@
   <ItemGroup>
     <Compile Include="Assets\Scripts\CameraController.cs" />
     <Compile Include="Assets\Scripts\CharacterController.cs" />
-    <Compile Include="Assets\Scripts\Enemy\EnemyController.cs" />
     <Compile Include="Assets\Scripts\GameEvents\Events\BaseGameEvent.cs" />
     <Compile Include="Assets\Scripts\GameEvents\Events\CharacterStatsEvent.cs" />
     <Compile Include="Assets\Scripts\GameEvents\Events\FloatEvent.cs" />
@@ -77,20 +76,19 @@
     <Compile Include="Assets\Scripts\GameEvents\UnityEvents\UnityFloatEvent.cs" />
     <Compile Include="Assets\Scripts\GameEvents\UnityEvents\UnityVoidEvent.cs" />
     <Compile Include="Assets\Scripts\GameEvents\Void.cs" />
-    <Compile Include="Assets\Scripts\ICharacterInput.cs" />
-    <Compile Include="Assets\Scripts\InputMethods\AIInput.cs" />
-    <Compile Include="Assets\Scripts\InputMethods\MouseAndKeyboardInput.cs" />
+    <Compile Include="Assets\Scripts\InputManager.cs" />
     <Compile Include="Assets\Scripts\Player\JetPack.cs" />
-    <Compile Include="Assets\Scripts\Player\PlayerController.cs" />
+    <Compile Include="Assets\Scripts\Player\PlayerMovement.cs" />
     <Compile Include="Assets\Scripts\PopupText.cs" />
-    <Compile Include="Assets\Scripts\RandomizeGroundTile.cs" />
-    <Compile Include="Assets\Scripts\RoomDestroyer.cs" />
-    <Compile Include="Assets\Scripts\RoomLists.cs" />
-    <Compile Include="Assets\Scripts\RoomSpawner.cs" />
+    <Compile Include="Assets\Scripts\Rooms&amp;Spawning\RandomizeGroundTile.cs" />
+    <Compile Include="Assets\Scripts\Rooms&amp;Spawning\RoomDestroyer.cs" />
+    <Compile Include="Assets\Scripts\Rooms&amp;Spawning\RoomLists.cs" />
+    <Compile Include="Assets\Scripts\Rooms&amp;Spawning\RoomSpawner.cs" />
+    <Compile Include="Assets\Scripts\Rooms&amp;Spawning\SpawnObjects.cs" />
+    <Compile Include="Assets\Scripts\Rooms&amp;Spawning\SpawnPlayer.cs" />
+    <Compile Include="Assets\Scripts\Rooms&amp;Spawning\TileSpriteList.cs" />
     <Compile Include="Assets\Scripts\ScriptableObjects\CharacterStats.cs" />
-    <Compile Include="Assets\Scripts\SpawnObjects.cs" />
-    <Compile Include="Assets\Scripts\SpawnPlayer.cs" />
-    <Compile Include="Assets\Scripts\TileSpriteList.cs" />
+    <Compile Include="Assets\Scripts\ScriptableObjects\Keybindings.cs" />
     <Compile Include="Assets\UnusedAssets\LevelGeneration.cs" />
     <Compile Include="Assets\UnusedAssets\Room.cs" />
     <Compile Include="Assets\UnusedAssets\SpanwRoom.cs" />
