+++ Assets/Resources/Prefab/JetPackFumesEffect.prefab
@@ -27,7 +27,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 0
+  m_IsActive: 1
 --- !u!4 &4177437680003186
 Transform:
   m_ObjectHideFlags: 1
