+++ Assets/Resources/Prefab/Tile.prefab
@@ -52,7 +52,7 @@ EdgeCollider2D:
   m_GameObject: {fileID: 1295147968759548}
   m_Enabled: 1
   m_Density: 1
-  m_Material: {fileID: 0}
+  m_Material: {fileID: 6200000, guid: fa2ca2a4888bdcb4380b05f03af46beb, type: 2}
   m_IsTrigger: 0
   m_UsedByEffector: 0
   m_UsedByComposite: 0
@@ -69,7 +69,7 @@ EdgeCollider2D:
   m_GameObject: {fileID: 1295147968759548}
   m_Enabled: 1
   m_Density: 1
-  m_Material: {fileID: 0}
+  m_Material: {fileID: 6200000, guid: fa2ca2a4888bdcb4380b05f03af46beb, type: 2}
   m_IsTrigger: 0
   m_UsedByEffector: 0
   m_UsedByComposite: 0
@@ -86,7 +86,7 @@ EdgeCollider2D:
   m_GameObject: {fileID: 1295147968759548}
   m_Enabled: 1
   m_Density: 1
-  m_Material: {fileID: 0}
+  m_Material: {fileID: 6200000, guid: fa2ca2a4888bdcb4380b05f03af46beb, type: 2}
   m_IsTrigger: 0
   m_UsedByEffector: 0
   m_UsedByComposite: 0
@@ -103,7 +103,7 @@ EdgeCollider2D:
   m_GameObject: {fileID: 1295147968759548}
   m_Enabled: 1
   m_Density: 1
-  m_Material: {fileID: 0}
+  m_Material: {fileID: 6200000, guid: fa2ca2a4888bdcb4380b05f03af46beb, type: 2}
   m_IsTrigger: 0
   m_UsedByEffector: 0
   m_UsedByComposite: 0
