--- Assets/ScriptableObjects/EnemyStats.asset
@@ -1,21 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInternal: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: ef2461f7c87a01d49885a17a645a4398, type: 3}
-  m_Name: EnemyStats
-  m_EditorClassIdentifier: 
-  HP: 50
-  MaxAscendVelocity: 0
-  MaxDecendVelocity: 0
-  MoveSpeed: 5
-  JumpForce: 0
-  IsFacingRight: 1
-  IsGrounded: 0
-  IsDead: 1
