+++ Assets/ScriptableObjects/InputMethods/Keyboard&Mouse.asset
@@ -1,0 +1,15 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: a86e9445aade95d47aca33e1dd7180f8, type: 3}
+  m_Name: Keyboard&Mouse
+  m_EditorClassIdentifier: 
+  jump: 32
+  attack: 323
