--- Assets/ScriptableObjects/PlayerStats.asset
@@ -1,22 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInternal: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: ef2461f7c87a01d49885a17a645a4398, type: 3}
-  m_Name: PlayerStats
-  m_EditorClassIdentifier: 
-  HP: 100
-  MaxAscendVelocity: 5
-  MaxDecendVelocity: 20
-  MoveSpeed: 5
-  SpeedSmoothTime: 0.1
-  JumpForce: 10
-  IsFacingRight: 0
-  IsGrounded: 1
-  IsDead: 0
