+++ Assets/ScriptableObjects/ScriptableObjectEvents/onPlayerJumped.asset
@@ -1,0 +1,13 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 280473b7f866f2242bdcb9c2b30d7ef6, type: 3}
+  m_Name: onPlayerJumped
+  m_EditorClassIdentifier:
