--- Assets/Scripts/ICharacterInput.cs
@@ -1,8 +1,0 @@
-﻿
-public interface ICharacterInput{
-    void ReadInput();
-    float Horizontal { get;   }
-    bool Attack { get;   }
-    bool Jump { get;  }
-    bool HeldJump { get; }
-}
