+++ Assets/Scripts/Rooms&Spawning/RoomDestroyer.cs
@@ -1,0 +1,12 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class RoomDestroyer : MonoBehaviour {
+
+    void OnTriggerEnter2D(Collider2D other)
+    {
+        if(other.GetComponent<RoomSpawner>())
+            Destroy(other.gameObject);
+    }
+}
