--- Assets/Scripts/SpawnObjects.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bfc833069d2e21f49bb3a33b202b3623
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
