--- Assets/Scripts/TileSpriteList.cs
@@ -1,12 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class TileSpriteList : MonoBehaviour {
-
-    public Sprite[] GroundTileSpriteListTop1;
-    public Sprite[] GroundTileSpriteListTop2;
-    public Sprite[] GroundTileSpriteListBottom;
-    public Sprite[] CeilingTileSpriteListAcid;
-    public Sprite GroundTile1Ledge;
-}
